requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
getkey==0.6.5
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
lxml==4.9.3
lyricsgenius==3.0.1
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
rapidfuzz==2.15.1
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
syncedlyrics==0.5.0
urllib3==1.26.16
